+++ inst/include/simmer/process/arrival_impl.h
@@ -1,5 +1,5 @@
 // Copyright (C) 2016 Bart Smeets and Iñaki Ucar
-// Copyright (C) 2016-2018 Iñaki Ucar
+// Copyright (C) 2016-2019 Iñaki Ucar
 //
 // This file is part of simmer.
 //
@@ -32,7 +32,7 @@ namespace simmer {
       return;
     }
 
-    foreach_ (ResMSet::value_type& itr, resources) {
+    foreach_ (ResVec::value_type& itr, resources) {
       Rcpp::warning("'%s': leaving without releasing '%s'", name, itr->name);
       itr->erase(this, true);
     }
@@ -58,11 +58,11 @@ namespace simmer {
       restime[ptr->name].start = sim->now();
       restime[ptr->name].activity = 0;
     }
-    resources.insert(ptr);
+    resources.push_back(ptr);
   }
 
   inline void Arrival::unregister_entity(Resource* ptr) {
-    ResMSet::iterator search = resources.find(ptr);
+    ResVec::iterator search = std::find(resources.begin(), resources.end(), ptr);
     if (!ptr || search == resources.end())
       Rcpp::stop("illegal unregister of arrival '%s'", name); // # nocov
     if (is_monitored())
@@ -70,23 +70,25 @@ namespace simmer {
     resources.erase(search);
   }
 
-  inline bool Arrival::leave_resources(bool flag) {
+  inline void Arrival::leave_resources(bool was_batched, bool keep_seized) {
     if (status.busy_until > sim->now())
       unset_busy(sim->now());
     unset_remaining();
-    while (resources.begin() != resources.end())
-      flag |= (*resources.begin())->erase(this);
-    return flag;
+    if (!keep_seized) {
+      while (resources.begin() != resources.end())
+        (*resources.begin())->erase(this);
+    } else if (!sim->is_scheduled(this) && !was_batched)
+      resources.back()->erase(this);
   }
 
-  inline void Arrival::renege(Activity* next) {
+  inline void Arrival::renege(Activity* next, bool keep_seized) {
+    bool in_batch = batch != NULL;
     timer = NULL;
     cancel_renege();
     if (batch && !batch->erase(this))
       return;
-    if (!leave_resources() && !batch)
-      deactivate();
-    batch = NULL;
+    leave_resources(in_batch, keep_seized);
+    deactivate();
     if (next) {
       activity = next;
       activate();