+++ tests/testthat/test-trajectory-renege.R
@@ -1,4 +1,4 @@
-# Copyright (C) 2016,2018 Iñaki Ucar
+# Copyright (C) 2016,2018-2019 Iñaki Ucar
 #
 # This file is part of simmer.
 #
@@ -525,4 +525,77 @@ test_that("seizes across nested batches are correctly reported", {
   expect_equal(arr_res$start_time, c(0, 0, 0))
   expect_equal(arr_res$end_time, c(2, 2, 2))
   expect_equal(arr_res$activity_time, c(2, 2, 2))
+})
+
+test_that("a reneging arrival keeps seized resources", {
+  out <- trajectory() %>% timeout(1)
+
+  t <- trajectory() %>%
+    renege_in(1, out=out, keep_seized=TRUE) %>%
+    seize("dummy") %>%
+    timeout(4)
+
+  env <- simmer(verbose = TRUE) %>%
+    add_resource("dummy") %>%
+    add_generator("arrival", t, at(0))
+
+  expect_warning(run(env))
+
+  arr <- get_mon_arrivals(env)
+  res <- get_mon_resources(env)
+
+  expect_equal(arr$end_time, 2)
+  expect_equal(arr$activity_time, 2)
+  expect_true(arr$finished)
+  expect_equal(res$time, 0)
+  expect_equal(res$server, 1)
+
+  t <- trajectory() %>%
+    renege_in(1, out=out, keep_seized=TRUE) %>%
+    seize("dummy") %>%
+    seize("other") %>%
+    timeout(4)
+
+  env <- simmer(verbose = TRUE) %>%
+    add_resource("dummy") %>%
+    add_resource("other", 0) %>%
+    add_generator("arrival", t, at(0))
+
+  expect_warning(run(env))
+
+  arr <- get_mon_arrivals(env)
+  res <- get_mon_resources(env)
+
+  expect_equal(arr$end_time, 2)
+  expect_equal(arr$activity_time, 1)
+  expect_true(arr$finished)
+  expect_equal(res$resource, c("dummy", "other", "other"))
+  expect_equal(res$time, c(0, 0, 1))
+  expect_equal(res$server, c(1, 0, 0))
+  expect_equal(res$queue, c(0, 1, 0))
+
+  t <- trajectory() %>%
+    renege_in(1, out=out, keep_seized=TRUE) %>%
+    seize("dummy") %>%
+    batch(1) %>%
+    seize("other") %>%
+    timeout(4)
+
+  env <- simmer(verbose = TRUE) %>%
+    add_resource("dummy") %>%
+    add_resource("other", 0) %>%
+    add_generator("arrival", t, at(0))
+
+  expect_warning(run(env))
+
+  arr <- get_mon_arrivals(env)
+  res <- get_mon_resources(env)
+
+  expect_equal(arr$end_time, 2)
+  expect_equal(arr$activity_time, 1)
+  expect_true(arr$finished)
+  expect_equal(res$resource, c("dummy", "other", "other"))
+  expect_equal(res$time, c(0, 0, 1))
+  expect_equal(res$server, c(1, 0, 0))
+  expect_equal(res$queue, c(0, 1, 0))
 })